moved ScenarioControls to scenario_controls.h, .cpp
