+++ src/ui/search_tabs.cpp
@@ -18,11 +18,3 @@ AstarTab::AstarTab(ScenarioService& i_scenario_service) : SearchTab(i_scenario_s
 
 }
 
-ScenarioControls::ScenarioControls(ScenarioService& i_scenario_service) : scenario_service(i_scenario_service) {
-  QVBoxLayout *controlsLayout = new QVBoxLayout{};
-  bucket_box = new QComboBox{};
-  controlsLayout->addWidget(bucket_box);
-  scenario_box = new QComboBox{};
-  controlsLayout->addWidget(scenario_box);
-  setLayout(controlsLayout);
-}
